requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# AI/ML and Language Models
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5

# Data Visualization
plotly>=5.17.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
coverage>=7.0.0

